api-recherche/requirements.txt
fastapi==0.104.1
Pillow==10.1.0
Pillow==10.1.0
pydantic==2.5.2
Requests==2.31.0
SQLAlchemy==2.0.23